Refactored the UI initialization to be outside of GameScreen
