+++ core/src/edu/sadsnails/game/Actions.java
@@ -104,7 +104,7 @@ public class Actions {
 		
 		System.out.println("Type: " + d_type + " \nSubject: " + d_subject);
 		
-		if(energy_consum >= state.energy) {
+		if(energy_consum >= state.getEnergy()) {
 			System.out.println("This task will take more energy than you have");
 			time_consum = 0;
 			energy_consum = 0;
@@ -188,7 +188,7 @@ public class Actions {
 			alterPopularity(-2);
 		}
 		incMoney(state.popularity, art_rank);
-		state.energy -= energy_consum;
+		state.setEnergy(state.getEnergy() - energy_consum);
 		incXP(xp_gain);
 		passTime(time_consum);
 		state.printStates();
@@ -270,7 +270,7 @@ public class Actions {
 	 */
 	public void sleep(int type) {		
 		// if energy is 100 you cannot sleep
-		if(state.energy == 100) {
+		if(state.getEnergy() == 100) {
 			System.out.println("You can't sleep when you're wide awake!");
 		}
 		// try to sleep
@@ -281,7 +281,7 @@ public class Actions {
 		if(type == 2) {
 			System.out.println("I successfully slept");
 			passTime(-1);
-			state.energy = 100;
+			state.setEnergy(100);
 		}
 
 		// if you have napped, you can't nap again
@@ -292,12 +292,12 @@ public class Actions {
 				System.out.println("I have already napped today");
 
 			else {
-				if(state.energy + 50 >= 100) {
-					state.energy = 100;
+				if(state.getEnergy() + 50 >= 100) {
+					state.setEnergy(100);
 					state.has_napped = true;
 				}
 				else {
-					state.energy += 50;
+					state.setEnergy(state.getEnergy() + 50);
 					state.has_napped = true;
 				}
 				passTime(5);
@@ -319,23 +319,23 @@ public class Actions {
 	public void buyBooster() {
 		if(state.coffee_used) 
 			System.out.println("I have already had coffee today");
-		else if(state.money < 5) {
+		else if(state.getMoney() < 5) {
 			System.out.println("I do not have enough money for this");
 		}
-		else if(state.energy == 100) {
+		else if(state.getEnergy() == 100) {
 			System.out.println("I cannot drink coffee when I am wide awake");
 		}
-		else if((state.energy + 40) > 100) {
+		else if((state.getEnergy() + 40) > 100) {
 			System.out.println("I bought coffee!");
-			state.energy = 100;
-			state.money -= 5;
+			state.setEnergy(100);
+			state.setMoney(state.getMoney() - 5);
 			state.spent_money += 5;
 			state.coffee_used = true;
 		}
 		else {
 			System.out.println("I bought coffee!");
-			state.energy += 40;
-			state.money -= 5;
+			state.setEnergy(state.getEnergy() + 40);
+			state.setMoney(state.getMoney() - 5);
 			state.spent_money += 5;
 			state.coffee_used = true;
 		}
@@ -347,38 +347,38 @@ public class Actions {
 	 */
 	public void incXP(int xpAmount) {
 		
-		if(xpAmount + state.xp >= state.toNext && state.level != 8) {
+		if(xpAmount + state.getXp() >= state.toNext && state.level != 8) {
 			//level up, if you're not max level
 			
-			state.xp	+= xpAmount;
+			state.setXp(state.getXp() + xpAmount);
 			state.level	++;
 			state.toNext *= 3;
 			
 			switch (state.level){
-			case 1: state.title = "1";
+			case 1: state.setTitle("1");
 					break;
-			case 2: state.title = "2";
+			case 2: state.setTitle("2");
 					break;
-			case 3: state.title = "3";
+			case 3: state.setTitle("3");
 					break;
-			case 4: state.title = "4";
+			case 4: state.setTitle("4");
 					break;
-			case 5: state.title = "5";
+			case 5: state.setTitle("5");
 					break;
-			case 6: state.title = "6";
+			case 6: state.setTitle("6");
 					break;
-			case 7: state.title = "7";
+			case 7: state.setTitle("7");
 					break;
-			case 8: state.title = "8";
+			case 8: state.setTitle("8");
 					break;
-			default: state.title = "0 WRONG";
+			default: state.setTitle("0 WRONG");
 					 break;
 			}
 			
-			System.out.println("You've leveled up! You are now level " +state.level + "! People have started to call you the " + state.title);
+			System.out.println("You've leveled up! You are now level " +state.level + "! People have started to call you the " + state.getTitle());
 		}
 		else
-			state.xp 	+= xpAmount;
+			state.setXp(state.getXp() + xpAmount);
 		}
 	
 	/* passTime method:
@@ -389,34 +389,34 @@ public class Actions {
 	 */
 	public void passTime(int hrs) {
 		if(hrs == -1) {
-			if(state.date[2] == 30) {
-				if(state.date[1] == 12) 
-					state.date[0] ++;
-				state.date[1] ++;
-				state.date[2] = 1;
+			if(state.getDate()[2] == 30) {
+				if(state.getDate()[1] == 12) 
+					state.getDate()[0] ++;
+				state.getDate()[1] ++;
+				state.getDate()[2] = 1;
 			}
 			else {
-				state.date[2] ++;
+				state.getDate()[2] ++;
 			}
-		 state.hour = 1;
+		 state.setHour(1);
 		}
-		else if((state.hour + hrs) > 24) {
-			int hr = 24 - state.hour;
-			if(state.date[2] == 30) {
-				if(state.date[1] == 12) 
-					state.date[0] ++;
-				state.date[1] ++;
-				state.date[2] = 1;
+		else if((state.getHour() + hrs) > 24) {
+			int hr = 24 - state.getHour();
+			if(state.getDate()[2] == 30) {
+				if(state.getDate()[1] == 12) 
+					state.getDate()[0] ++;
+				state.getDate()[1] ++;
+				state.getDate()[2] = 1;
 			}
 			else {
-				state.date[2] ++;
+				state.getDate()[2] ++;
 			}
-			state.hour = hr + hrs;
+			state.setHour(hr + hrs);
 			state.has_napped = false;
 			state.coffee_used = false;
 		}
 		
-		state.hour += hrs;
+		state.setHour(state.getHour() + hrs);
 		
 	}
 	
@@ -431,23 +431,23 @@ public class Actions {
 		if(popularity == -1) {
 			switch(ranking) {
 			case -2:
-				state.money += 0;
+				state.setMoney(state.getMoney() + 0);
 				state.earned_money += 0;
 				break;
 			case -1:
-				state.money += 0;
+				state.setMoney(state.getMoney() + 0);
 				state.earned_money += 0;
 				break;
 			case 0:
-				state.money += 1;
+				state.setMoney(state.getMoney() + 1);
 				state.earned_money += 1;
 				break;
 			case 1:
-				state.money += 5;
+				state.setMoney(state.getMoney() + 5);
 				state.earned_money += 5;
 				break;
 			case 2:
-				state.money += 10;
+				state.setMoney(state.getMoney() + 10);
 				state.earned_money += 10;
 				break;
 			}
@@ -455,23 +455,23 @@ public class Actions {
 		if(popularity == -0.5) {
 			switch(ranking) {
 			case -2:
-				state.money += 0;
+				state.setMoney(state.getMoney() + 0);
 				state.earned_money += 0;
 				break;
 			case -1:
-				state.money += 1;
+				state.setMoney(state.getMoney() + 1);
 				state.earned_money += 1;
 				break;
 			case 0:
-				state.money += 2;
+				state.setMoney(state.getMoney() + 2);
 				state.earned_money += 2;
 				break;
 			case 1:
-				state.money += 7;
+				state.setMoney(state.getMoney() + 7);
 				state.earned_money += 7;
 				break;
 			case 2:
-				state.money += 8;
+				state.setMoney(state.getMoney() + 8);
 				state.earned_money += 8;
 				break;
 			}
@@ -479,23 +479,23 @@ public class Actions {
 		if(popularity == 0) {
 			switch(ranking) {
 			case -2:
-				state.money += 1;
+				state.setMoney(state.getMoney() + 1);
 				state.earned_money += 1;
 				break;
 			case -1:
-				state.money += 5;
+				state.setMoney(state.getMoney() + 5);
 				state.earned_money += 5;
 				break;
 			case 0:
-				state.money += 10;
+				state.setMoney(state.getMoney() + 10);
 				state.earned_money += 10;
 				break;
 			case 1:
-				state.money += 12;
+				state.setMoney(state.getMoney() + 12);
 				state.earned_money += 12;
 				break;
 			case 2:
-				state.money += 15;
+				state.setMoney(state.getMoney() + 15);
 				state.earned_money += 15;
 				break;
 			}
@@ -503,23 +503,23 @@ public class Actions {
 		if(popularity == 0.5) {
 			switch(ranking) {
 			case -2:
-				state.money += 2;
+				state.setMoney(state.getMoney() + 2);
 				state.earned_money += 2;
 				break;
 			case -1:
-				state.money += 5;
+				state.setMoney(state.getMoney() + 5);
 				state.earned_money += 5;
 				break;
 			case 0:
-				state.money += 15;
+				state.setMoney(state.getMoney() + 15);
 				state.earned_money += 15;
 				break;
 			case 1:
-				state.money += 20;
+				state.setMoney(state.getMoney() + 20);
 				state.earned_money += 20;
 				break;
 			case 2:
-				state.money += 25;
+				state.setMoney(state.getMoney() + 25);
 				state.earned_money += 25;
 				break;
 			}
@@ -527,23 +527,23 @@ public class Actions {
 		if(popularity == 1) {
 			switch(ranking) {
 			case -2:
-				state.money += 3;
+				state.setMoney(state.getMoney() + 3);
 				state.earned_money += 3;
 				break;
 			case -1:
-				state.money += 6;
+				state.setMoney(state.getMoney() + 6);
 				state.earned_money += 6;
 				break;
 			case 0:
-				state.money += 20;
+				state.setMoney(state.getMoney() + 20);
 				state.earned_money += 20;
 				break;
 			case 1:
-				state.money += 25;
+				state.setMoney(state.getMoney() + 25);
 				state.earned_money += 25;
 				break;
 			case 2:
-				state.money += 30;
+				state.setMoney(state.getMoney() + 30);
 				state.earned_money += 30;
 				break;
 			}
